+++ src/panoptes/pocs/utils/cli/main.py
@@ -3,8 +3,9 @@ import typer
 from panoptes.pocs.utils.cli import config
 from panoptes.pocs.utils.cli import sensor
 from panoptes.pocs.utils.cli import image
-from panoptes.pocs.utils.cli import power
+from panoptes.pocs.utils.cli import mount
 from panoptes.pocs.utils.cli import notebook
+from panoptes.pocs.utils.cli import power
 from panoptes.pocs.utils.cli import run
 from rich import print
 
@@ -13,6 +14,7 @@ state = {'verbose': False}
 
 app.add_typer(config.app, name="config", help='Interact with the config server.')
 app.add_typer(image.app, name="image", help='Interact with images.')
+app.add_typer(mount.app, name="notebook", help='Simple mount controls.')
 app.add_typer(notebook.app, name="notebook", help='Start Jupyter notebook environment.')
 app.add_typer(power.app, name="power", help='Interact with power relays.')
 app.add_typer(run.app, name="run", help='Run POCS!')
